update script with interval

+++ examples/run-homeassistant.sh
@@ -1,5 +1,12 @@
 #!/usr/bin/env bash
 
-rm nohup.out
+if [ -f nohup.out ]; then
+  rm nohup.out
+fi
 
-nohup python3 mqtt-all-homeassistant.py --broker emqx.home-assistant.localdomain &
+BROKER="emqx.home-assistant.localdomain"
+INTERVAL=60
+
+python3 mqtt-all-homeassistant.py --broker ${BROKER} --interval ${INTERVAL}
+
+# nohup python3 mqtt-all-homeassistant.py --broker emqx.home-assistant.localdomain &
